+++ pkg/kubernetes/api/sloth/v1/types.go
@@ -1,8 +1,6 @@
 package v1
 
 import (
-	"encoding/json"
-
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -58,7 +56,7 @@ type SLO struct {
 	// +kubebuilder:validation:Required
 	//
 	// Objective is target of the SLO the percentage (0, 100] (e.g 99.9).
-	Objective json.Number `json:"objective"`
+	Objective float64 `json:"objective"`
 
 	// Labels are the Prometheus labels that will have all the recording and
 	// alerting rules for this specific SLO. These labels are merged with the